basic error handling added

+++ controllers/articles.js
@@ -1,3 +1,4 @@
+import mongoose from 'mongoose'
 import express from 'express'
 import Article from '../models/Article.js'
 
@@ -6,32 +7,73 @@ const router = express.Router()
 // ! Routes that render a web page
 // Index - get all articles
 router.get('/articles', async (req, res) => {
-  const allArticles = await Article.find()
-  return res.render('articles/index.ejs', {
-    articles: allArticles
-  })
+  try {
+    const allArticles = await Article.find()
+    return res.render('articles/index.ejs', {
+      articles: allArticles
+    })
+  } catch (error) {
+    console.log(error)
+  }
 })
 
 // New - Displays the form that allows us to submit a CREATE request
 router.get('/articles/new', (req, res) => {
-  return res.render('articles/new.ejs')
+  try {
+    return res.render('articles/new.ejs', {
+      errorMessage: ''
+    })
+  } catch (error) {
+    console.log(error)
+  }
 })
 
 // Edit - Displays the form that allows us to submit an UPDATE request
-router.get('/articles/:articleId/edit', async (req, res) => {
-  const article = await Article.findById(req.params.articleId)
-  console.log(article)
-  return res.render('articles/edit.ejs', {
-    article
-  })
+router.get('/articles/:articleId/edit', async (req, res, next) => {
+  try {
+    // Validate incoming articleId
+    if (!mongoose.isValidObjectId(req.params.articleId)){
+      return next()
+    }
+
+    // Search the DB for the article with the provided ID
+    const article = await Article.findById(req.params.articleId)
+
+    // If article isn't found, return 404 by running next()
+    if (!article) return next()
+
+    // If article was found, render the page
+    return res.render('articles/edit.ejs', {
+      article
+    })
+  } catch (error) {
+    console.log(error)
+  }
 })
 
 // Show - displays a single article
-router.get('/articles/:articleId', async (req, res) => {
-  const article = await Article.findById(req.params.articleId)
-  return res.render('articles/show.ejs', {
-    article: article
-  })
+router.get('/articles/:articleId', async (req, res, next) => {
+  try {
+    // Before we use the article to find the article
+    // Lets check the id is valid
+    // If it is NOT valid, we'll send an error response
+    if (!mongoose.isValidObjectId(req.params.articleId)){
+      return next()
+    }
+
+    // Using the id to find an article
+    const article = await Article.findById(req.params.articleId)
+
+    // If findById fails to find a matching article to the id provided, it will return null
+    // If it returns null, we want to send a 404, by running next()
+    if (!article) return next()
+
+    return res.render('articles/show.ejs', {
+      article: article
+    })
+  } catch (error) {
+    console.log(error)
+  }
 })
 
 
@@ -44,7 +86,10 @@ router.post('/articles', async (req, res) => {
     const newArticle = await Article.create(req.body)
     return res.redirect(`/articles/${newArticle._id}`)
   } catch (error) {
-    console.log(error)
+    console.log(error.message)
+    return res.render('articles/new.ejs', {
+      errorMessage: error.message
+    })
   }
 })
 
@@ -52,7 +97,19 @@ router.post('/articles', async (req, res) => {
 router.put('/articles/:articleId', async (req, res) => {
   try {
     const articleId = req.params.articleId
-    await Article.findByIdAndUpdate(articleId, req.body)
+
+    // Validate incoming articleId
+    if (!mongoose.isValidObjectId(articleId)){
+      return next()
+    }
+
+    // Attempt to make the update for the article
+    const updatedArticle = await Article.findByIdAndUpdate(articleId, req.body)
+
+    // If article not found, return 404
+    if (!updatedArticle) return next()
+
+
     return res.redirect(`/articles/${articleId}`)
   } catch (error) {
     console.log(error)
@@ -61,8 +118,15 @@ router.put('/articles/:articleId', async (req, res) => {
 
 // Delete - allows us to delete an existing article
 router.delete('/articles/:articleId', async (req, res) => {
-  await Article.findByIdAndDelete(req.params.articleId)
-  return res.redirect('/articles')
+  try {
+    if (!mongoose.isValidObjectId(req.params.articleId)){
+      return next()
+    }
+    await Article.findByIdAndDelete(req.params.articleId)
+    return res.redirect('/articles')
+  } catch (error) {
+    console.log(error)
+  }
 })
 
 // ! Don't forget to export your router
